+++ subprojects/core/src/test/groovy/org/gradle/api/internal/file/archive/TarFileTreeTest.java
@@ -23,7 +23,7 @@ import org.gradle.api.resources.MissingResourceException;
 import org.gradle.api.resources.internal.LocalResourceAdapter;
 import org.gradle.test.fixtures.file.TestFile;
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider;
-import org.gradle.util.Resources;
+import org.gradle.util.internal.Resources;
 import org.junit.Rule;
 import org.junit.Test;
